requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
daal==2021.2.3
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.110.1
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.13.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
g4f==0.2.8.1
gradio==4.8.0
gradio_client==0.7.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
openai==0.28.0
openai-whisper==20231117
orjson==3.10.0
packaging==24.0
pandas==2.2.1
pillow==10.3.0
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
PyExecJS==1.5.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.5
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.11.4
semantic-version==2.10.0
sentence-transformers==2.6.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
tbb==2021.10.0
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
tqdm==4.66.2
transformers==4.39.3
typer==0.12.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
websockets==11.0.3
yarl==1.9.4
zipp==3.18.1
